AM335X: CM3: Enable the MBX interrupt for recovery
During a debug session it was discovered that for some
reason M3 and A8 are expecting different interrupts to
be handled. This should not have happened in the first place
and needs further analysis. For now we keep the MBX interrupt
enabled so that A8 can issue a reset command to M3.

While here, a trivial error in function declaration which
the compiler didn't catch earlier has been fixed.

M3 firmware version has been incremented to help differentiate
binaries with and without the MBX interrupt fix.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>