AM335X: CM3: Add CORE LDO retention feature for PG 2.0
PG 2.0 supports CORE LDO retention which will help in
lowering DeepSleep0 power consumption. CORE LDO retention
should be enabled only when the OSC is going to be shut
down and hence an appropriate is used.

Note: There shouldn't be any check for the PER domain in
retention before CORE LDO retention is enabled since the
status bit in the PRCM register is meant for this purpose.
However, we are being smart in trying to leverage the DS0
sequence for USB wakeup which should be used only in DS1
mode. One of the ways for avoiding issues with USB wakeup
in pseudo DS1 (DS0 forced to DS1 by leaving a peripheral
active) is to put an additional check for the PER domain
status before the CORE LDO is programmed to enter retention.
The other approach is to make the kernel DS1 aware which
is not what we want and hence we go with the extra smart
solution ;)

Also bump up the M3 version.

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>
[vaibhav.bedia@ti.com: Update the patch to take care of
OSC dependency, put the CORE LDO retention under the
appropriate check for PER domain and upate the commit
message for the same]
Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>