AM335x: CM3: Update the DS2 resume path to issue interrupt
As part of the DS2 entry MPU executes the WFI instruction.
When an async wakeup event happens, not all peripherals
will issue an interrupt to the MPU INTC which will make
the MPU come out of WFI state. The right solution here
is to have the M3 issue an interrupt to MPU as part of
the DS2 resume process.

At the same time we also update the M3 version and binary.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>