AM335X: CM3: Fix DPLL power up issue for PG 2.0 in DS1/2
Commit 345ff63 (AM335X: CM3: Add DPLL retention feature for PG 2.0)
missed up adding the steps to power up the DPLLs in DS1/2 modes.
This results in the DPLLs never powering up and hence hanging the
system in the resume from DS1/2 modes. Fix this by adding the
required power up sequence. Also update the firmware version and
the pre-built binaries to reflect this change.

Special thanks to Dandamudi, Ramesh <x0182916@ti.com> for debugging
the issue and finding the right fix.

Signed-off-by: Dandamudi, Ramesh <x0182916@ti.com>
Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>